Fix bug where quicksort test wasn't showing quicksort output
The test code would always show the result of mergesort when doing the
verbose test of quicksort.  As a result, if quicksort was broken, the
verbose test of quicksort would still display a correctly sorted list.

This fix affects only the test code that runs when `sorts.py` is run as
a script.  It doesn't change anything in the sorting functions
themselves.